perf!: fix an O(n^5) meltdown, and drop f64 for i32/f32 throughout
Two changes, both prompted by claims in this repo that turned out to be
untrue when measured.

## The meltdown

A benchmark was the first thing missing. Adding one showed the crate did
not scale at O(n^2 log n) as the docs claimed, but at n^5.5: a
132-vertex comb took 124ms, and a 260-vertex one asked for 27GB and died.
The claim was theory; nobody had run it.

The cause was a dependency, not a constant. A split event was stamped
with the two endpoints of the edge it was aimed at, so a vertex's event
depended on a part of the wavefront arbitrarily far away. Any event
invalidated events all over the polygon; every reschedule registered more
dependencies; that made the next event invalidate even more. It fed back
on itself.

The fix is not better bookkeeping but a better question. A split's
*timing* never depended on those endpoints: an edge's wavefront slides
along its own offset track and never leaves it, so the moment a vertex
meets that track is fixed by the vertex's trajectory and the edge's
original line, and nothing elsewhere can change it. The endpoints only
decide whether the landing is on a live stretch -- and that is settled
exactly when the event is popped, because events come off in time order,
so now == t and the wavefront's shape is fact rather than forecast.

So split_lower_bound queues a deliberately weak answer: the first time
the vertex reaches any edge's line, live or not. It can be early, never
late, which is all that popping in time order needs. handle_split_event
then looks rather than guesses. Events now depend on O(1) vertices -- an
owner and its neighbour -- and the dependents index is gone entirely.

A smaller cause hid behind it: the struck-off list was scanned linearly
inside the per-edge loop, so one scan cost O(n * rejections). Sorting it
and binary-searching took a random star from n^3.0 to n^2.2.

Measured at 1024 vertices: convex 1.1ms (~n^1.3, genuinely sub-quadratic
-- no reflex vertices means no split search at all), comb 8.8ms (~n^1.9),
random star 13ms (~n^2.1). The event count is linear; the quadratic term
is the O(n) scan per reflex vertex, inherent to this formulation. Beating
it needs the motorcycle graph -- a different, much larger algorithm --
and the docs say so rather than claim otherwise.

## No f64

The brief asked for i32/f32 so the algorithm could run where f64 is slow
or absent. The crate used f64 internally and i64 in its predicates,
because i16 inputs need 35 bits for an orientation determinant. True, but
solving the wrong problem: a type you only use "internally" is still a
type the hardware has to have.

The answer is one bit of coordinate range. Coordinates are capped at
-16384..=16383, and Polygon rejects anything wider. One expression sets
the cap -- the orientation determinant, needing 2*d^2 for the largest
coordinate difference d:

    full i16:  2 * 65535^2 = 8_589_672_450  -> overflows i32, wrapping
                                               to the *wrong side*
    capped:    2 * 32767^2 = 2_147_352_578  -> fits, 131_069 to spare

So one bit buys an exact i32 predicate: no epsilon, no rounding, no
overflow. And it is exactly one bit -- the capped worst case uses 99.994%
of i32.

f32 cannot do that job at any range. A search found a triple *inside* the
cap -- (14176,-12146), (-9937,5341), (4434,-5081) -- whose true
determinant is 9 where f32 returns 0, calling a genuine turn collinear.
That is now a test, beside the out-of-range triple where i32 flips sign.
Both were found by search; an earlier draft asserted an f32 failure that
turned out not to be real.

The simulation is f32 throughout, and the cap is what leaves it enough
absolute resolution: 0.002 at the worst corner, against 0.004 uncapped.
The docs are straight about the cost -- f64 resolved ~1e-11 there, and
adversarially near-degenerate input can be closer than f32 can see. What
makes that survivable is that robustness here is structural rather than
numeric: needle zipping and node dedup are tolerance-based by design. If
the cap ever bites, double-float (a pair of f32s, ~48 mantissa bits) is
the honest next step, and the docs say so.

ring_area2 keeps an i64 accumulator and is documented as the exception:
summing triangles can wind around a region more than once, so the total
is bounded by vertex count rather than the coordinate box. It is
validation -- once, on the host, never in the simulation.

All 160 tests pass, including every geometric invariant and a new one at
the coordinate cap, where f32 has least room.

BREAKING CHANGE: coordinates are capped at -16384..=16383. Polygon now
returns PolygonError::CoordinateOutOfRange for anything wider. Scale
input down to fit.

+++ src/roof.rs
@@ -433,8 +433,8 @@ fn lattice_height(h: f32) -> Option<i16> {
     if !h.is_finite() {
         return None;
     }
-    let r = round_half_away_from_zero(h as f64);
-    if r < i16::MIN as f64 || r > i16::MAX as f64 {
+    let r = round_half_away_from_zero(h);
+    if r < i16::MIN as f32 || r > i16::MAX as f32 {
         return None;
     }
     Some(r as i16)
